Initialize splits array if data not found in project altas
If a 9patch asset is deleted from the project, and then a new asset
with the same name is added Hyperlap2D will crash. This crash continues
to occur if the project is opened back up, meaning the altas file needs
to be manually edited to fix the project. This commit inits the splits
array if the splits data could not be found in the project atlas, to
prevent the application from continuously crashing when loading the
project.